backend_python/app.py
from flask import Flask, request, jsonify
import os
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

app = Flask(__name__)

@app.route('/')
def hello():
    return "Python backend is running!"

@app.route('/api/speech-to-text/realtime', methods=['POST'])
def realtime_speech_to_text():
    # Placeholder for real-time speech recognition logic
    # This will involve receiving audio stream and sending to Google Cloud Speech-to-Text
    return jsonify({"message": "Real-time endpoint hit"}), 200

@app.route('/api/speech-to-text/asynchronous', methods=['POST'])
def asynchronous_speech_to_text():
    # Placeholder for asynchronous speech recognition logic
    # This will involve receiving an audio file, uploading to GCS (optional),
    # and starting a long-running operation with Google Cloud Speech-to-Text
    if 'audio_file' not in request.files:
        return jsonify({"error": "No audio file part"}), 400
    file = request.files['audio_file']
    if file.filename == '':
        return jsonify({"error": "No selected file"}), 400
    if file:
        # Process file here (e.g., save it, send to GCS, etc.)
        return jsonify({"message": f"File {file.filename} received. Asynchronous processing started."}), 202

# Placeholder for authentication routes
@app.route('/api/auth/login', methods=['POST'])
def login():
    return jsonify({"message": "Login endpoint placeholder"}), 200

@app.route('/api/auth/register', methods=['POST'])
def register():
    return jsonify({"message": "Register endpoint placeholder"}), 200

# Placeholder for database related routes
@app.route('/api/data', methods=['GET'])
def get_data():
    return jsonify({"message": "Get data endpoint placeholder"}), 200

@app.route('/api/data', methods=['POST'])
def post_data():
    return jsonify({"message": "Post data endpoint placeholder"}), 200


if __name__ == '__main__':
    port = int(os.environ.get('PORT', 8080))
    app.run(debug=True, host='0.0.0.0', port=port) 