backend_python/requirements.txt
Flask
python-dotenv
google-cloud-speech
opencv-contrib-python==4.8.0.74
tqdm==4.65.0
matplotlib==3.7.2
numpy==1.25.1
torch==2.1.2
click==8.1.6
requests==2.31.0
mediapipe==0.10.11
deep_translator==1.11.4
